[OBMIN-326] Create  ResetAllButton
